requirements.txt
requests
pycryptodome
protobuf
urllib3